Fix some compilation warning.
git-svn-id: http://svn.modules.centreon.com/centreon-broker/trunk@955 7c7d615a-03b3-410d-a316-1135b1bd9499

+++ core/src/logging/manager.cc
@@ -87,7 +87,7 @@ void manager::_compute_optimizations() {
          end = _backends.end();
        it != end;
        ++it)
-    for (unsigned int i = 1; i <= it->l; ++i)
+    for (unsigned int i = 1; i <= static_cast<unsigned int>(it->l); ++i)
       _limits[i] |= it->types;
   return ;
 }
@@ -206,7 +206,9 @@ void manager::log_on(backend& b,
     p.l = min_priority;
     p.types = types;
     _backends.push_back(p);
-    for (unsigned int i = 1; i <= min_priority; ++i)
+    for (unsigned int i = 1;
+         i <= static_cast<unsigned int>(min_priority);
+         ++i)
       _limits[i] |= types;
     connect(
       &b,
